Separate `ControllerRegistration` from `Admission`
Earlier, the code differentiated virtual and runtime resources which became confusing due to the amount of functions and terms. This commit rather splits it at `ControllerRegistration`s (and `ControllerDeployment`s) and `Admission` (and in the future `Extension`, deployed in the runtime cluster).
